+++ R/s3-sparse.R
@@ -3,7 +3,7 @@
 #  sparsebnUtils
 #
 #  Created by Bryon Aragam (local) on 1/22/16.
-#  Copyright (c) 2014-2016 Bryon Aragam. All rights reserved.
+#  Copyright (c) 2014-2017 Bryon Aragam. All rights reserved.
 #
 
 #------------------------------------------------------------------------------#
@@ -147,33 +147,15 @@ sparse.list <- function(x, ...){
 #
 #' @export
 sparse.matrix <- function(x, index = "R", ...){
-    if( nrow(x) != ncol(x)) stop("Input matrix must be square!") # 2-7-15: Why does it need to be square?
+    matrix_to_sparse(x, index = "R", ...)
+} # END SPARSE.MATRIX
 
-    if(index != "R" && index != "C") stop("Invalid entry for index parameter: Must be either 'R' or 'C'!")
-
-    pp <- nrow(x)
-
-    nnz <- which(abs(x) > zero_threshold()) - 1
-    vals <- double(length(nnz))
-    rows <- integer(length(nnz))
-    cols <- integer(length(nnz))
-
-    x <- as.vector(x)
-    for(k in seq_along(nnz)){
-        col <- trunc(nnz[k] / pp)
-        row <- nnz[k] - (pp * col)
-        vals[k] <- x[nnz[k] + 1]
-        rows[k] <- row
-        cols[k] <- col
-    }
-
-    sp <- sparse.list(list(rows = as.integer(rows), cols = as.integer(cols), vals = as.numeric(vals), dim = c(pp, pp), start = 0))
-
-    if(index == "R"){
-        reIndexR(sp)
-    } else{
-        sp
-    }
+#------------------------------------------------------------------------------#
+# sparse.Matrix
+#
+#' @export
+sparse.Matrix <- function(x, index = "R", ...){
+    Matrix_to_sparse(x, index = "R", ...)
 } # END SPARSE.MATRIX
 
 #' @export
@@ -194,25 +176,6 @@ sparse.edgeList <- function(x, ...){
 
     sparse(out)
 }
-
-# #------------------------------------------------------------------------------#
-# # as.sparse.list
-# #  Convert FROM list TO sparse
-# #
-# #' @export
-# as.sparse.list <- function(x, ...){
-#     sparse.list(x)
-# } # END AS.SPARSE.LIST
-#
-# #------------------------------------------------------------------------------#
-# # as.sparse.matrix
-# #  Convert FROM matrix TO sparse
-# #  By default, return the object using R indexing. If desired, the method can return C-style indexing by setting
-# #    index = "C".
-# #' @export
-# as.sparse.matrix <- function(x, index = "R", ...){
-#     sparse.matrix(x, index)
-# } # END AS.SPARSE.MATRIX
 
 #------------------------------------------------------------------------------#
 # as.matrix.sparse
@@ -280,14 +243,120 @@ is.zero.sparse <- function(x){
 } # END IS.ZERO.SPARSE
 
 #------------------------------------------------------------------------------#
-# .num_edges.sparse
+# num.nodes.sparse
+#
+#' @export
+num.nodes.sparse <- function(x){
+    x$dim[2]
+} # NUM.NODES.SPARSE
+
+#------------------------------------------------------------------------------#
+# num.edges.sparse
+#
+#' @export
+num.edges.sparse <- function(x){
+    ### What to do about this special case...
+    # length(x$rows) # Ignores potentially very small edge weights which may be zero
+
+    .num_edges(x)
+} # NUM.EDGES.SPARSE
+
+#------------------------------------------------------------------------------#
+# t.sparse
+#  Take implicit transpose by swapping rows <-> cols
+#
+#' @export
+t.sparse <- function(x){
+    temp <- x$rows
+    x$rows <- x$cols
+    x$cols <- temp
+
+    x
+}
+
+#------------------------------------------------------------------------------#
+# .num_edges
 # Internal function for returning the number of edges in a sparse object
 #
-.num_edges.sparse <- function(x){
-    ### Testing only for now
-    if(length(which(abs(x$vals) > zero_threshold())) != length(x$rows)){
-        stop("Error in .num_edges.sparse! Please check source code.")
+.num_edges <- function(x, threshold = FALSE){
+    stopifnot(is.sparse(x))
+
+    if(!threshold){
+        length(x$rows)
+    } else{
+        ### Testing only for now
+        if(length(which(abs(x$vals) > zero_threshold())) != length(x$rows)){
+            stop("Error in .num_edges.sparse! Please check source code.")
+        }
+
+        length(which(abs(x$vals) > zero_threshold()))
     }
 
-    length(which(abs(x$vals) > zero_threshold()))
-} # END .NUM_EDGES.SPARSE
+} # END .NUM_EDGES
+
+matrix_to_sparse <- function(x, index = "R", ...){
+    stopifnot(check_if_matrix(x))
+
+    if( nrow(x) != ncol(x)) stop("Input matrix must be square!") # 2-7-15: Why does it need to be square?
+
+    if(index != "R" && index != "C") stop("Invalid entry for index parameter: Must be either 'R' or 'C'!")
+
+    pp <- nrow(x)
+
+    # t1 <- proc.time()[3] ################################################
+    nnz <- Matrix::which(is.na(x) | (abs(x) > zero_threshold()), arr.ind = TRUE)
+    # t2 <- proc.time()[3] ################################################
+    # cat(sprintf("as.sparse nnz via which(...): %f\n", t2-t1))
+
+    ### This is a weird hack that is surprisingly fast:
+    ###  Note that the output of which is in the same order as as.vector,
+    ###  in the sense that the first element in as.vector(x[x != 0])
+    ###  corresponds exactly to the first row of nnz (which is the output
+    ###  of which).
+    ###
+    ### This is _substantially_ faster than either Matrix::Matrix
+    ###  and an Rcpp implementation of the loop that was previously used.
+    vals <- as.vector(x[x != 0])
+    rows <- nnz[, 1]
+    cols <- nnz[, 2]
+
+    sp <- sparse.list(list(rows = as.integer(rows), cols = as.integer(cols), vals = as.numeric(vals),
+                           dim = c(pp, pp),
+                           start = 1
+                           )
+                      )
+
+    if(index == "R"){
+        suppressWarnings(reIndexR(sp))
+    } else{
+        suppressWarnings(reIndexC(sp))
+    }
+} # END MATRIX_TO_SPARSE
+
+Matrix_to_sparse <- function(x, index = "R", ...){
+    stopifnot(check_if_matrix(x))
+
+    if( nrow(x) != ncol(x)) stop("Input matrix must be square!") # 2-7-15: Why does it need to be square?
+
+    if(index != "R" && index != "C") stop("Invalid entry for index parameter: Must be either 'R' or 'C'!")
+
+    pp <- nrow(x)
+
+    ### If already a Matrix object, just pull out the corresponding slots
+    Tx <- as(x, "dgTMatrix") # needs to be in triplet form, see ?dgCMatrix vs ?dgTMatrix
+    vals <- Tx@x
+    rows <- Tx@i
+    cols <- Tx@j
+
+    sp <- sparse.list(list(rows = as.integer(rows), cols = as.integer(cols), vals = as.numeric(vals),
+                           dim = c(pp, pp),
+                           start = 0 # NOTE: dgTMatrix uses C-style indexing, so this is different from matrix_to_sparse!
+                           )
+                      )
+
+    if(index == "R"){
+        suppressWarnings(reIndexR(sp))
+    } else{
+        suppressWarnings(reIndexC(sp))
+    }
+} # END MATRIX_TO_SPARSE