+++ R/s3-sparsebnData.R
@@ -3,7 +3,7 @@
 #  sparsebnUtils
 #
 #  Created by Bryon Aragam (local) on 1/28/16.
-#  Copyright (c) 2014-2016 Bryon Aragam. All rights reserved.
+#  Copyright (c) 2014-2017 Bryon Aragam. All rights reserved.
 #
 
 #------------------------------------------------------------------------------#
@@ -95,8 +95,6 @@
 #'
 #' @examples
 #'
-#' \dontrun{
-#'
 #' ### Generate a random continuous dataset
 #' mat <- matrix(rnorm(1000), nrow = 20)
 #' dat <- sparsebnData(mat, type = "continuous") # purely observational data with continuous variables
@@ -115,8 +113,6 @@
 #'                     type = "discrete",
 #'                     levels = dat.levels,
 #'                     ivn = dat.ivn) # specify intervention rows
-#'
-#' }
 #'
 #' @docType class
 #' @name sparsebnData