+++ tests/testthat/test-edgeList.R
@@ -14,6 +14,18 @@ test_that("edgeList constructor checks for inconsistent indices", {
     expect_error(edgeList(list(1, 1, c(1, -1))))    # -1 is not a valid index
 })
 
+test_that("Permuting nodes results in isomorphic output", {
+    ### NOTE: These tests are random, but should return TRUE for any input, so randomization is not an issue
+    gr <- random.graph(3,3)
+    expect_true(igraph::isomorphic(to_igraph(gr), to_igraph(permute.nodes(gr))))
+
+    gr <- random.graph(10,20)
+    expect_true(igraph::isomorphic(to_igraph(gr), to_igraph(permute.nodes(gr))))
+
+    gr <- random.graph(50,200)
+    expect_true(igraph::isomorphic(to_igraph(gr), to_igraph(permute.nodes(gr))))
+})
+
 test_that("Bugfix: edgeList does not accept NULL or NA values", {
     expect_error(edgeList(list(NULL, NULL)), "Input cannot have any NULL values!")
     expect_error(edgeList(list(1:2, NULL, 2)), "Input cannot have any NULL values!")